--- src/Formats/NullFormat.cpp
@@ -1,23 +1,0 @@
-#include <Formats/NullFormat.h>
-#include <Formats/FormatFactory.h>
-#include <IO/WriteBuffer.h>
-
-
-namespace DB
-{
-
-WriteBuffer NullOutputFormat::empty_buffer(nullptr, 0);
-
-void registerOutputFormatNull(FormatFactory & factory)
-{
-    factory.registerOutputFormat("Null", [](
-        WriteBuffer &,
-        const Block & sample,
-        const RowOutputFormatParams &,
-        const FormatSettings &)
-    {
-        return std::make_shared<NullOutputFormat>(sample);
-    });
-}
-
-}
